pyproject.toml
[tool.poetry]
name = "stanforteedge"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "mit"

[tool.poetry.scripts]
server = "stanforteedge.server:run_app"


[tool.poetry.dependencies]
python = "^3.7"
starlette = "^0.20.0"
uvicorn = "^0.17.6"
tortoise-orm = {extras = ["aiosqlite"], version = "^0.19.1"}
Jinja2 = "^3.1.2"
install = "^1.3.5"
Starlette-WTF = "^0.4.3"
aiosqlite = "^0.17.0"


[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
